refactor: split app.ts in two files
